Moved sym unit tests
