+++ crates/sym/src/lib.rs
@@ -9,3 +9,6 @@ pub use crate::convert::concretize;
 pub use crate::convert::concretize_into;
 pub use crate::convert::ConcreteValue;
 pub use crate::sym::*;
+
+#[cfg(test)]
+mod tests;
