+++ crates/sym/src/tests.rs
@@ -1,0 +1,6 @@
+mod aiger;
+mod bit;
+mod bitbuf;
+mod bitvec;
+mod convert;
+mod pcode;
