+++ crates/sym/src/tests/pcode.rs
@@ -1,0 +1,7 @@
+use crate::*;
+use pcode_ops::validate;
+
+#[test]
+fn validate_pcode() -> validate::Result {
+    validate::Validator::<SymbolicBitVec>::validate()
+}
